# [performance] Avoid linear walk of graph children.