--- a/src/dag/dag.ml
+++ b/src/dag/dag.ml
@@ -12,12 +12,16 @@ module Make (Value : Value) : S with type value := Value.t = struct
 
     type graph = t
 
+    module Node_map = Map.Make(Int)
+
     type node_info =
       { id : int
       ; (* only used for printing *)
         mutable mark : mark
       ; mutable level : int
       ; mutable deps : node list
+      ; (* see #2959, we need to implement is_child efficiently *)
+        mutable deps_set : unit Node_map.t
       ; mutable rev_deps : node list
       ; mutable parent : node option
       }
@@ -60,7 +64,8 @@ module Make (Value : Value) : S with type value := Value.t = struct
     let set_parent _ v p = v.info.parent <- Some p
 
     let raw_add_edge _ v w =
-      v.info.deps <- w :: v.info.deps
+      v.info.deps <- w :: v.info.deps;
+      v.info.deps_set <- Node_map.add_exn v.info.deps_set w.info.id ()
 
     let raw_add_vertex _ _ = ()
   end
@@ -75,7 +80,7 @@ module Make (Value : Value) : S with type value := Value.t = struct
   let create_node_info g =
     let id = g.fresh_id in
     g.fresh_id <- g.fresh_id + 1;
-    { id; mark = -1; level = 1; deps = []; rev_deps = []; parent = None }
+    { id; mark = -1; level = 1; deps = []; deps_set = Node_map.empty; rev_deps = []; parent = None }
 
   let add g v w =
     match IC.add_edge_or_detect_cycle g v w with
@@ -102,6 +107,5 @@ module Make (Value : Value) : S with type value := Value.t = struct
 
   let pp_node pp_value fmt n = pp_depth 0 pp_value fmt n
 
-  let is_child v w =
-    v.info.deps |> List.exists ~f:(fun c -> c.info.id = w.info.id)
+  let is_child v w = Node_map.mem v.info.deps_set w.info.id
 end