--- a/src/dag/dag.ml
+++ b/src/dag/dag.ml
@@ -2,10 +2,9 @@ open! Stdune
 include Dag_intf
 
 module Make (Value : Value) : S with type value := Value.t = struct
-
-  (* Raw_graph here should have the same complexity than the assumed
-     interface on the incremental_cycles proofs, in particular
-     [get_outgoing] should run in constant time. *)
+  (* Raw_graph here should have the same complexity than the assumed interface
+     on the incremental_cycles proofs, in particular [get_outgoing] should run
+     in constant time. *)
   module Raw_graph = struct
     type mark = int
 
@@ -16,7 +15,7 @@ module Make (Value : Value) : S with type value := Value.t = struct
 
     type graph = t
 
-    module Node_map = Map.Make(Int)
+    module Node_map = Map.Make (Int)
 
     type node_info =
       { id : int
@@ -84,7 +83,14 @@ module Make (Value : Value) : S with type value := Value.t = struct
   let create_node_info g =
     let id = g.fresh_id in
     g.fresh_id <- g.fresh_id + 1;
-    { id; mark = -1; level = 1; deps = []; deps_set = Node_map.empty; rev_deps = []; parent = None }
+    { id
+    ; mark = -1
+    ; level = 1
+    ; deps = []
+    ; deps_set = Node_map.empty
+    ; rev_deps = []
+    ; parent = None
+    }
 
   let add g v w =
     match IC.add_edge_or_detect_cycle g v w with
@@ -114,10 +120,12 @@ module Make (Value : Value) : S with type value := Value.t = struct
   let is_child v w = Node_map.mem v.info.deps_set w.info.id
 
   let add_idempotent g v w =
-    (* if the edge doesn't already exist, we
-       add it to the graph; note that the complexity guarantees for
-       `Dag.add` don't hold if the edge is already in the graph,
-       hence the check , see #2959 for more details and the
-       README of the vendored library *)
-    if is_child v w then () else add g v w
+    (* if the edge doesn't already exist, we add it to the graph; note that the
+       complexity guarantees for `Dag.add` don't hold if the edge is already in
+       the graph, hence the check , see #2959 for more details and the README of
+       the vendored library *)
+    if is_child v w then
+      ()
+    else
+      add g v w
 end