--- a/src/dag/dag.ml
+++ b/src/dag/dag.ml
@@ -92,9 +92,9 @@ module Make (Value : Value) : S with type value := Value.t = struct
     ; parent = None
     }
 
-  (* [add_assuming_missing dag v w] creates an arc going from [v] to [w]. @raise Cycle if
-      creating the arc would create a cycle. This assumes that the arc does not
-      already exist. *)
+  (* [add_assuming_missing dag v w] creates an arc going from [v] to [w]. @raise
+     Cycle if creating the arc would create a cycle. This assumes that the arc
+     does not already exist. *)
   let add_assuming_missing g v w =
     match IC.add_edge_or_detect_cycle g v w with
     | IC.EdgeAdded -> ()