When calling Dune in scenarios where targets have a large number of
deps, Dune will take a long time to start. A common case is when
depending on `(package coq)`, which brings into the DAG a few thousand
files as children of a single node.

`perf` data show this is due to the linear walk in `Dag.is_child`;
indeed, doing a naive replacement of the list for a more efficient
access structure solves the problem. For example a dummy target that depends on `(package coq)` does:

```
with this PR:

real	0m1,684s
user	0m1,552s
sys	0m0,128s

with master:

real	0m11,450s
user	0m10,587s
sys	0m0,264s
```

This PR is just a proof-of-concept and it must not be merged yet; I've
opened it to track the issue and to devise what would the best
solution be.

As far as I can see there are several options, I am not familiar
enough with the cycle detection algo as to propose something now.