@ejgallego 2019-12-09
@corwin-of-amber when this is ready it should also resolve the bad build latency you were suffering in jscoq.