@diml 2019-12-11
@ejgallego is there more to do on this PR? Appart from updating the tests.