@ejgallego 2019-12-11
> @ejgallego is there more to do on this PR? Appart from updating the tests.

I am not sure the current fix is the right one; it was just done as a POC.

Note that we are casting from the map to a list often so this PR as-is could introduce performance regressions in other areas.

IMHO these castings should be avoided, I can take care of that but it will take me some time as I need to understand well that part of the code; feel free to take over.