@aalekseyev 2019-12-11
I thought that maybe this doesn't affect asymptotic time complexity if the call sites are already linear-time. Well, it's not clear that they're linear time: the list returned by `get_outgoing` is used in `interruptible_fold` which interrupts after expending some amount of "fuel".

An easy way to get rid of this potential complexity increase is to keep both the list and the map.