@ejgallego 2019-12-11
> An easy way to get rid of this potential complexity increase is to keep both the list and the map.

That's what I was thinking as a first fix; but I didn't go that route as maybe it would be worth to look into this with more detail and avoid such duplication if possible.