@diml 2019-12-12
What about implementing `interruptible_fold` with a `Map.fold` and `raise_notrace Exit`?