@aalekseyev 2019-12-12
@diml, yeah, that should work.
(by the way, a paranoid concern is that the asymptotic analysis might rely on the order and we're changing the order, but I'm happy to say that seems unlikely)