@diml 2019-12-12
BTW, `incremental_cycles` is a vendored library. I suggest to submit a PR upstream and discuss this with the author.