@ejgallego 2020-02-04
Hi @Armael , after some discussion we were thinking of changing the type of `get_outgoing` to an id-indexed map instead of a list; that would mean the `interruptible_fold` in https://github.com/ocaml/dune/blob/79a27e50dbd440ce0348d24174fb3cb8a0492ec3/vendor/incremental-cycles/src/incremental_cycles.ml#L156 would be a fold over a map [instead of the current list, which is walked in reverse addition order]

@Armael , do you think that would matter for `incremental cycles`? Is the order of the walk important? Thanks!