@ejgallego 2020-02-04
Ok folks, I've reworked the patch; this could be closer to something we could merge. Main point is maybe what to do with `get_deps_from_graph_exn` (cc: @snowleopard ) , also @Armael may confirm that by changing the fold in `incremental_cycles` we are not breaking any invariant.

I've reproduced the perf numbers on my own setup [building jsCoq] , would be interesting to understand better the possible performance implications if anyone has a large project.