@Armael 2020-02-04
Thanks for pinging me.

This discussion is in fact related to a shortcoming of the interface provided by `incremental cycles`, that is on my todo-list of things to improve, but that I haven't addressed yet.

More concretely, the *current* specification for the algorithm requires the `get_outgoing` function provided by the client to return a list of all successors, and do so *in constant time*. This is quite demanding: this basically requires the client to already have the list at hand.

Would `get_outgoing` have a complexity linear in the output list, then we would get an incorrect asymptotic complexity for the overall algorithm. This is because the algorithm might need to explore only a part of the list; and constructing the complete list beforehand might be already too expensive. (in particular in the case of a node with many children.)

What I want to implement, and which I believe is the best solution in the long run, is to require `get_outgoing` to return an iterator (`'a Seq.t`). This allows both the producer of the iterator and `interruptible_fold` to be implemented efficiently.

I'll have a look, but it will probably take me a bit of time to get through the proofs.

In the short term, I can see 3 options:

- Having `get_outgoing` return a map would work for this particular application; this is quite ad-hoc, and doesn't solve the general issue: `get_outgoing` would still be tied to a specific return datatype.
- Construct and maintain lists of successors along the map, in the graph datatype. This allows you to produce the list in constant time, and satisfies the existing interface for `incremental cycles`. The downside is the extra bookkeeping that is require to make sure the map and the list stay in sync.
- Implement the `Seq.t` solution. If you want, you can temporarily do the modification in dune (including in the vendored code), with the hope that I eventually prove that it is correct and re-synchronize the vendored code with the upstream.