@Armael 2020-02-04
A variation on point 3: you could submit the implementation using `Seq.t` as a PR on the upstream repo, targeting a feature branch. Then use that branch as vendoring target in dune.