@ejgallego 2020-02-04
Hi @Armael , thanks a lot for your quick turnaround,

> More concretely, the _current_ specification for the algorithm requires the `get_outgoing` function provided by the client to return a list of all successors, and do so _in constant time_.

I think this is (fortunately) the case in dune master and in this PR, we already have the list of successors at hand, the performance problem happens in a different path due to the way we store the successors.

> What I want to implement, and which I believe is the best solution in the long run, is to require `get_outgoing` to return an iterator (`'a Seq.t`). This allows both the producer of the iterator and `interruptible_fold` to be implemented efficiently.

It seems the main client is indeed `interruptible_fold`, so I'm going to try to provide an abstract interface for that; I think that's better than `Seq.t`, as it introduces a bit more of overhead that it is not needed for quite a few datatypes; the clients could use indeed to implement the interruptible fold using `Seq.t` instead.