@Armael 2020-02-04
> I think this is (fortunately) the case in dune master and in this PR, we already have the list of successors at hand, the performance problem happens in a different path due to the way we store the successors.

Ah, you do? Currently, the PR seems to be using `NodeMap.values` which I suspect is linear.

>  It seems the main client is indeed interruptible_fold, so I'm going to try to provide an abstract interface for that; I think that's better than Seq.t, as it introduces a bit more of overhead that it is not needed for quite a few datatypes; the clients could use indeed to implement the interruptible fold using Seq.t instead.

Ah, I see, we could indeed simply represent the sequence by its elimination principle (!), i.e. `interruptible_fold`. A few comments:
- for consistency, we need to mirror whatever change we do for `get_outgoing` in `get_incoming` as well (you would get the same issue if you decided to stop representing rev deps as lists);
- I'm unsure whether using `interruptible_fold` as specification instead of `Seq.t` is actually better. It makes the specification very higher-order (of order 3), i.e. the *client* has to provide a function that adheres to a higher-order specification (of order 2). Thinking how to specify side-effects, etc. 
That's not to say that specifying a `Seq.t` is easy, though François Pottier has some previous work about that (cf http://gallium.inria.fr/~fpottier/publis/fpottier-hashtable.pdf). But we should thought about it.
- If you were thinking about performance, is `Seq.t` really inefficient? It might be worth getting some actual measurements. 
- I think it would be a bit nicer if we could move the discussion about API changes in `incremental_cycles` to its own repo, and do the changes there first (you can do a PR targeting a feature branch; we don't need to think about proofs right away).