@Armael 2020-02-04
(if you're reluctant using incremental cycles' repo because it's on inria's gitlab, then we could also discuss moving it to github. I don't mind doing that, it's just a bit annoying.)